requirements.txt
# Fabric CLI CI/CD - Production Dependencies
# Dev/test tools are in requirements-dev.txt

# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
jsonschema>=4.19.0
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
gitpython>=3.1.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-keyvault-secrets>=4.7.0

# Gap closing enhancements
pydantic>=2.5.0
pydantic-settings>=2.1.0
jinja2>=3.1.2
packaging>=23.0
